+++ funcs_aux/__init__.py
@@ -1,3 +1,3 @@
 from .np import array_2d_get_compact_str
 from .collects import Iterables
-from .results import ResultSucceed
+from .results import ResultSucceedSimple