+++ funcs_aux/results.py
@@ -1,17 +1,23 @@
-from typing import NamedTuple, Any
+from typing import *
 
 
 # =====================================================================================================================
-class ResultSucceed(NamedTuple):
+class ResultSucceedSimple(NamedTuple):
     """
-    main idea - solve NONE-value ambiguity
-    by simple way
+    MAIN GOAL:
+    ----------
+    solve NONE-value as result ambiguity by simple way
+
+    MAIN RULES:
+    -----------
+    return object if you get final result!
+    return None if there are any errors in execution
 
     USAGE:
     ------
-    def func(a, b) -> Optional(ResultSucceed):
+    def func(a, b) -> Optional(ResultSucceedSimple):
         if a in b:
-            return ResultSucceed(a)
+            return ResultSucceedSimple(a)
         else:
             return
 
@@ -19,7 +25,7 @@ class ResultSucceed(NamedTuple):
     assert result is None
 
     result = func(None, [None, ])
-    assert result == ResultSucceed(None)
+    assert result == ResultSucceedSimple(None)
 
     if result:
         print(result.VALUE)
@@ -30,5 +36,50 @@ class ResultSucceed(NamedTuple):
     def __call__(self, *args, **kwargs) -> Any:
         return self.VALUE
 
+
+# =====================================================================================================================
+class ResultFull:
+    """
+    MAIN GOAL:
+    ----------
+    use always Object as result with absolut unambiguity!
+
+    MAIN RULES:
+    -----------
+    return object in any cases (if you get final result or if there are any errors in execution)!
+    """
+    # TODO: add TESTS
+    # TODO: add TESTS
+    # TODO: add TESTS
+    # TODO: add TESTS
+    # TODO: add TESTS
+    # TODO: add TESTS
+
+    # INPUT
+    FUNC: Callable
+    ARGS: Tuple[Any, ...] = []
+    KWARGS: Dict[str, Any] = {}
+
+    # RESULTS
+    OK: Optional[bool] = None
+    RESULT: Optional[Any] = None
+    EXX: Optional[Exception] = None
+
+    def __call__(self, *args, **kwargs) -> None:
+        # init ---------------
+        self.OK = None
+        self.RESULT = None
+        self.EXX = None
+
+        self.ARGS = args        # FIXME: resolve work with args!
+        self.KWARGS = kwargs    # FIXME:
+
+        # WORK ---------------
+        try:
+            self.RESULT = self.FUNC(*args, **kwargs)
+            self.OK = True
+        except Exception as exx:
+            self.EXX = exx
+            self.OK = False
 
 # =====================================================================================================================