+++ test__result.py
@@ -1,14 +1,14 @@
-from funcs_aux import ResultSucceed
+from funcs_aux import ResultSucceedSimple
 
 
 # =====================================================================================================================
 def test__ResultSucceed():
-    assert ResultSucceed(123)() == 123
+    assert ResultSucceedSimple(123)() == 123
 
-    assert ResultSucceed(123).VALUE == 123
-    assert ResultSucceed([123]).VALUE == [123]
-    assert ResultSucceed({123}).VALUE == {123}
-    assert ResultSucceed({123: 123}).VALUE == {123: 123}
+    assert ResultSucceedSimple(123).VALUE == 123
+    assert ResultSucceedSimple([123]).VALUE == [123]
+    assert ResultSucceedSimple({123}).VALUE == {123}
+    assert ResultSucceedSimple({123: 123}).VALUE == {123: 123}
 
 
 # =====================================================================================================================